% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DirectingOptiRounds.R
\name{DirectingOptiRounds}
\alias{DirectingOptiRounds}
\title{Directing Optimization Round of a SCF Convergence Round in the Tibble Formed in OptiSCFMonitorAsWholeTibble()}
\usage{
DirectingOptiRounds(Pending, index)
}
\arguments{
\item{Pending}{A integer refering to the row of the directed SCF convergence cycle is in.}

\item{index}{A tibble including all the row number}
}
\value{
A integer directing the Optimization round of that SCF convergence round is in.
}
\description{
This function is a internal function that directs the optimization round of a SCF convergence round. This helps SCFMonitor devide the SCF rounds to different sections that each refers to an optimization cycle.
}
\examples{
library(tibble)

example_index <- tibble(rowid = c(1, 30, 130))
DirectingOptiRounds(33, example_index)

}
