% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormOptiSCFConvergenceRoundTibble.R
\name{FormOptiSCFConvergenceRoundTibble}
\alias{FormOptiSCFConvergenceRoundTibble}
\title{Form a Tibble of Total SCF Rounds in the Optimization Cycles}
\arguments{
\item{directory}{A string vector describing the directory of the Gaussian log file.}
}
\value{
A tibble countain two columns, describing each optimization rounds and the number of SCF rounds it undergoes until convergence.
}
\description{
This function outputs a tibble showing each the number of SCF rounds applied in each optimization cycle and outputs it as a tibble.
}
\examples{
library(readr)
library(stringr)
library(tidyselect)
library(dplyr)
library(tibble)

FormOptiSCFConvergenceRoundTibble(SCFMonitorExample())

}
