% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptiSCFConvergenceRoundMonitor.R
\name{OptiSCFConvergenceRoundMonitor}
\alias{OptiSCFConvergenceRoundMonitor}
\title{Read and Plot the Counts the SCF Convergence Rounds of Each Optimization Step of a Gaussian Log File.}
\arguments{
\item{directory}{A string vector describing the directory of the Gaussian log file.}
}
\value{
No return value, called for side effects
}
\description{
This function reads a log file automatically and generate a plot showing the steps it takes to reach SCF convergence for each optimization process.
}
\examples{
library(ggplot2)

OptiSCFConvergenceRoundMonitor(SCFMonitorExample())

}
