% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptiSCFMonitorAsList.R
\name{OptiSCFMonitorAsList}
\alias{OptiSCFMonitorAsList}
\title{Form a list Containing SCF Data of Each Optimization Cycles}
\arguments{
\item{directory}{A string vector describing the directory of the Gaussian log file.}
}
\value{
A list of lists. First lists is a list of tibble, each element in the list refers to a tibble recording the SCF Data of a optimization step. The second list only have one element that is a numeric vector refering to the SCF convergence requirement read from log file.
}
\description{
This function reads a Gaussian .log file and outputs a list of tibbles, each of which is the SCF Data of a optimization step.
}
\examples{
library(readr)
library(stringr)
library(tidyselect)
library(utils)
library(dplyr)
library(tibble)

OptiSCFMonitorAsList(SCFMonitorExample())

}
