% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptiSCFMonitorAsWholeTibble.R
\name{OptiSCFMonitorAsWholeTibble}
\alias{OptiSCFMonitorAsWholeTibble}
\title{Form a Tibble of SCF data for each Optimization Steps}
\arguments{
\item{directory}{A string vector describing the directory of the Gaussian log file.}
}
\value{
A list containing two elements. The first one is a tibble containing the SCF data of every rounds labeled with the optimization steps they are in. The second element is a numeric vector that refers to the SCF convergence standard.
}
\description{
This function outputs a tibble containing the data of each rounds of SCF calculation labeled with the optimization round it's in (if it's a optimization job, otherwise it will be only 1)
}
\examples{
library(readr)
library(stringr)
library(tidyselect)
library(utils)
library(dplyr)
library(tibble)

OptiSCFMonitorAsWholeTibble(SCFMonitorExample())

}
