% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCFMonitorExample.R
\name{SCFMonitorExample}
\alias{SCFMonitorExample}
\title{Acquire Path to SCFMonitor Example data}
\usage{
SCFMonitorExample()
}
\value{
A string vector showing the path to
}
\description{
Makes the TestData.log, a log file generated by Gaussian 16 for testing the package, easy to access.
}
\examples{
FormOptiSCFConvergenceRoundTibble(SCFMonitorExample())

}
